third_party/blink/web_tests/fast/dom/shadow/anchor-content-projected-expected.txt

CONSOLE WARNING: line 27: Element.createShadowRoot is deprecated and will be removed in M80, around February 2020. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 and https://developers.google.com/web/updates/2019/07/web-components-time-to-upgrade for more details.
Links should be clickable even when their content is projected into them.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Cursor Info: type=Hand hotSpot=0,0
PASS window.location.hash is "#link-clicked"
PASS successfullyParsed is true

TEST COMPLETE

Example link
